Rename trie/polkadot_trie_db to trie/impl; Refactor trie codec
