+++ core/storage/trie/impl/CMakeLists.txt
@@ -1,22 +1,28 @@
 # Copyright Soramitsu Co., Ltd. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-add_library(polkadot_trie_db_codec
+add_library(polkadot_trie_codec
     polkadot_codec.cpp
-    polkadot_node.cpp
     )
-target_link_libraries(polkadot_trie_db_codec
+target_link_libraries(polkadot_trie_codec
     buffer
     blake2
     blob
     scale
     )
 
+add_library(polkadot_node
+    polkadot_node.cpp
+    )
+target_link_libraries(polkadot_node
+    buffer
+    )
+
 add_library(polkadot_trie_db
     polkadot_trie_db.cpp
     )
 target_link_libraries(polkadot_trie_db
-    polkadot_trie_db_codec
+    polkadot_trie_codec
     polkadot_trie
     )
 
@@ -26,6 +32,7 @@ add_library(polkadot_trie
 target_link_libraries(polkadot_trie
     buffer
     polkadot_trie_batch
+    polkadot_node
     )
 
 add_library(polkadot_trie_batch
@@ -39,6 +46,6 @@ add_library(ordered_trie_hash INTERFACE)
 
 target_link_libraries(ordered_trie_hash INTERFACE
     polkadot_trie
-    polkadot_trie_db_codec
+    polkadot_trie_codec
     scale
     )
