+++ core/storage/trie/impl/polkadot_trie_db.cpp
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "storage/trie/polkadot_trie_db/polkadot_trie_db.hpp"
+#include "storage/trie/impl/polkadot_trie_db.hpp"
 
 #include <utility>
 
-#include "storage/trie/polkadot_trie_db/polkadot_codec.hpp"
-#include "storage/trie/polkadot_trie_db/polkadot_node.hpp"
-#include "storage/trie/polkadot_trie_db/polkadot_trie.hpp"
-#include "storage/trie/polkadot_trie_db/polkadot_trie_batch.hpp"
+#include "storage/trie/impl/polkadot_codec.hpp"
+#include "storage/trie/impl/polkadot_node.hpp"
+#include "storage/trie/impl/polkadot_trie.hpp"
+#include "storage/trie/impl/polkadot_trie_batch.hpp"
 
 using kagome::common::Buffer;
