+++ core/storage/trie/impl/polkadot_trie_db.hpp
@@ -10,9 +10,9 @@
 #include <optional>
 
 #include "crypto/hasher.hpp"
-#include "storage/trie/polkadot_trie_db/polkadot_codec.hpp"
-#include "storage/trie/polkadot_trie_db/polkadot_node.hpp"
-#include "storage/trie/polkadot_trie_db/polkadot_trie.hpp"
+#include "storage/trie/impl/polkadot_codec.hpp"
+#include "storage/trie/impl/polkadot_node.hpp"
+#include "storage/trie/impl/polkadot_trie.hpp"
 #include "storage/trie/trie_db.hpp"
 
 namespace kagome::storage::trie {
