+++ test/core/storage/trie/trie_db/polkadot_codec_hash256_test.cpp
@@ -7,7 +7,7 @@
 
 #include <gtest/gtest.h>
 #include "crypto/blake2/blake2b.h"
-#include "storage/trie/polkadot_trie_db/polkadot_codec.hpp"
+#include "storage/trie/impl/polkadot_codec.hpp"
 
 using namespace kagome;
 using namespace common;
