+++ test/core/storage/trie/trie_db/polkadot_codec_node_decoding_test.cpp
@@ -6,9 +6,9 @@
 #include <memory>
 
 #include <gtest/gtest.h>
-#include "storage/trie/polkadot_trie_db/polkadot_codec.hpp"
-#include "storage/trie/polkadot_trie_db/polkadot_node.hpp"
-#include "storage/trie/polkadot_trie_db/buffer_stream.hpp"
+#include "storage/trie/impl/polkadot_codec.hpp"
+#include "storage/trie/impl/polkadot_node.hpp"
+#include "storage/trie/impl/buffer_stream.hpp"
 #include "testutil/outcome.hpp"
 #include "testutil/literals.hpp"
